GB: Significantly improved PPU timings.
* Turns out the old implementation used WAY too many dot clocks per frame.
* Added CPU/refresh rate to the debug window.
* Added a ton of asserts and debug stuff.

+++ libplip/PlipInstance.cpp
@@ -67,7 +67,7 @@ Plip::PlipError PlipInstance::Load(const PlipValidCore core, const std::string &
     return m_core->Load(path);
 }
 
-void PlipInstance::Run(const long ns) const {
+void PlipInstance::Run(const double ns) const {
     m_core->Delta(ns);
 }
